Try different min-cut stage orders. When cutting the graph into stages in the pipeline schedule, try different ordering of which stages to cut. Previously, the cutting was done in ascending order (cut graph for pipeline registers after stage 0, then after stage 1, ...., until the last stage). With this CL ascending, descending, and middle-first are also attempted and the schedule with the minimum number of registers is kept.
Middle-first means the cut in the middle of the pipeline (stage ~n/2 for an n stage pipeline) is done first. This divides the graph in two pieces (before and after stage n/2). Each piece is then cut in the middle in a recursive fashion until the graph has been cut into the desired n pieces.

This heuristic tempers the effect of greediness of the cutting. Specifically, it is possible that a min-cut for stage X could result in less optimal cut for a later cut for stage Y, while doing it in a reverse order (Y then X) could have a better result.

PiperOrigin-RevId: 313622638