Add Project Icestorm Bazel dependency to XLS
This is a dependency of nextpnr. I chose to not include the iceprog program, because it depends on libftdi, which in turn depends on a bunch of other things. I'm not sure it's worth pulling these things in at this time. It can be done later if/when somebody actually wants it.

PiperOrigin-RevId: 314357517