Implements per-pipeline-stage LEC.
This change enables an individual pipeline stage to be LEC'ed (between source IR and generated netlist) by extracting the stage from each source - the stage inputs are replaced by a common set of Z3 constant nodes. The rest of the LEC proceeds as normal - the outputs from each implementation are checked for equality (after "unflattening" the netlist output).

PiperOrigin-RevId: 314985859