Use a hermetic C/C++ toolchain (LLVM 10) when building XLS
This ensures that everyone who works with XLS gets the same set of compiler warnings and that the build works with the same set of compiler flags. In the future it's likely that the build system will be expanded to work with more compilers but for now having to support several compilers takes effort and doesn't really help anybody.

This uses a toolchain specification published by a biotech company grail.com, but the change should be safe from a security point of view because the configuration downloads official LLVM binaries from official mirrors. The code provided by Grail is all in source form.

Note: This change triggers quite a few warnings compared to when building with GCC 9.
PiperOrigin-RevId: 315291295