Allow rebinding multiple netlist inputs at the same time.
Rebinding an input into a netlist "tree" is not a cheap operation - it requires invoking Z3_substitute on any downstream nets, which could be a very large set. In real-world examples, the necessary time can be measured in seconds. To reduce this cost, this change enables _all_ netlist inputs to be rebound at once, avoiding the need for repeated tree traversals.

While at it, I greatly simplified the tree traversal logic and possibly fixed a bug or two in it.

PiperOrigin-RevId: 316561553