Add array simplification optimization pass. Currently the pass optimizes ArrayIndex operations with literal indices when the element at that index can be determined at compile time
Background: when the ArrayUpdate operation was added to the IR and tied to the DSLX operation update(), the optimized IR generated from std::find_index regressed in some cases. Specifically, array update and array index operations remain in the IR after optimizations while previously (before use of arrayupdate) no such array operations remained.

For better testing of these kinds of problems, the test xls/tests/dslx_optimization_test.cc is added which tests the optimized IR which result from DSLX constructs. It is in C++ because the complicated part of the test is the matching of the IR which is a C++ api. Currently it only tests a specific form of std::find_index.

PiperOrigin-RevId: 316591030