Extends the IR JIT to operate upon bit-packed views.
This change adds LlvmIrJit::RunWithPackedViews(), which accepts arguments as a list of input view types followed by a single output view type. Internally, the JIT sends these into LLVM-space as a list of pointers to iX types, where X is the packed size of each argument. When referencing each parameter, the iX is loaded from the input buffer and exploded, element-by-element, into a regular LLVM type.

On program completion, the same process is followed in reverse - the result LLVM type is blitted, element-by-element, into an iX type, which is then copied into the result view's buffer.

Part of OSS REPO issue #58.

PiperOrigin-RevId: 327325535