Plumb AST owner through all AST types.
This prepares for the switch to a pybind11 C++ AST, since ownership in C++
relates to an "owning" object's lifetime. You could hide many of these for AST
nodes that are operations on AST node operands, but this makes the API uniform
in that you know you need access to the ownership object when you create a new
AST node (at the cost of making it a bit more verbose in the common case).

PiperOrigin-RevId: 328344872