Switch name from NodeToType to TypeInfo.
It used to be simply a mapping from AST node to its concretized type, but now
it has more complexity (deals with imports, can be instantiated with a parent
when concretizing parametric functions, etc.)

I'm also going to hoist more things onto to it
(e.g. the maps mutably held on Invocation nodes).

PiperOrigin-RevId: 329322688