Eliminating the now-unnecessary "high_cells" netlist/LEC tool flag.
Our Liberty file support has evolved to the point where we no longer need to hardcode certain cells to certain values (outside the usual constraint specification functionality), so we can drop this flag.

Although our Liberty support is adequate for our purposes, it is not _complete_, and will likely receive changes in the future. For now, though, all necessary functionality is present.

PiperOrigin-RevId: 330531138