Try to use v1 cache for CI and v2 cache for nightly fuzz.
Nightly fuzz is triggered by schedule and so cannot use the v1 cache. I'm not
sure if they will work together so this is trial by fire.

Also switched home directory to use `~` as a path element since I see this
being done in documentation.

Caching doesn't seem to be working for pushes to main, so trying to switch
things up and see if it gets better.

PiperOrigin-RevId: 331658112