Port DSLX parser to C++
Unfortunately this is the rote transliteration of 1.4k of
Python code to about 2k of C++ code, so it's a pretty massive
change. This is the downside of the "prototype in Python
and when ideas are validated to a decent degree port to C++"
approach.

Porting DSL infra to C++ before adding more support for
concurrent processes to avoid making the port even larger.

PiperOrigin-RevId: 332125594