Adding basic "table switch" optimization pass.
This pass identifies constructs of the conceptual form:

```
if (a == 0) {
  x = y0;
} else if (a == 1) {
  x == y1;
} else if (a == 2) {
  x == y2;
}
```

and so on, and converts them into an lookup in a constant array, e.g.,

```
int foo[] = {y0, y1, y2, ...};
```

This first-pass implementation is limited to patterns where the index increments or decrements by 1, where either the first or last comparison is 0, and where there are no "gaps" in the index space (but subspaces without gaps will be transformed: indexes [0, 1, 2, 5, 6, 7] will result in two table lookups.

Much further optimization is possible:
 - non-zero start or end indexes
 - non-unit strides,
 - "gap" handling
and likely more, but this is enough code/complexity to review and land for now.

PiperOrigin-RevId: 334240179