Adds new IR -> SMTLIB2 converter.
Part of this (really, the bulk of the work) is adding the ability to "booleanify" the input IR, which is to say, to decompose IR ops into their fundamenal AND/OR/NOT operations. The SMTLIB2 conversion can then be done on that decomposed IR (the higher-level IR can instead be converted, of course).

PiperOrigin-RevId: 335919568