Add unique id to IR node text representation. The unique ID is emitted as an attribute like so:
   foobar: bits[32] = add(x, y, id=42)

This change is necessary because we can no longer rely on the name to carry the unique ID.

The vast majority of this change is patching up change-detector tests which string match IR text. The C++ tests are all migrated to IR matcher. The python tests remain (passing for now) change detectors because we have no good way of matching the IR in python.

PiperOrigin-RevId: 336093575