[DSLX] Port TypeInfo to C++
Also upgrades pybind11 to a version with fixed rvalue reference py::cast'ing.

Note that this exposes some danger via the Python API because we need to break
cycles among TypeInfo. That is, people can leak if they don't call `clear_type_info_refs_for_gc()`, because the TypeInfos can refer to each other,
and there are reflexive 'parent' vs 'imported' links.

An ImportCache helper object is
provided for the common use case where you've got a main driver
program that is the lifetime root for all the modules.

No easy way to expose pybind11 object fields to the cycle
collector that I can see, but since this will all be ported over to C++ with
simplified API entry points after this salvo of changes is done, it seems
reasonable.

PiperOrigin-RevId: 336144237