Fix isinstance() assertion for binding the parameters of a struct.
Parameters can be u32, or can be functions of other parameters. This isinstance() check was meant to check for references to other parameters, but uses the wrong class. Now it uses the right one.

A test is also added. The test uses parametric structs in a for expression, which triggers this behavior.

PiperOrigin-RevId: 336793521