Logical equivalence updates/fixes.
A) Add ReduceOr op to Z3IrTranslator.
B) Fix a bug in Z3IrTranslator::HandleSMul: If an operation overflowed the initial operand sizes, then data could be lost if the result size was larger (than the operands). Simple fix - expand the operands as wide as possible before doing the operation, then shrink afterwards if necessary.

PiperOrigin-RevId: 339770254