Add streaming and single-value channel types. A streaming channel has FIFO semantics and can hold an arbitrary number of elements. A single-value channel holds, well, a single value. A receive on a single-value channel returns the last element sent to the channel.
A single-value channel might represent a register or a module port. A single-value channel is fundamentally a lower level of abstraction than a streaming channel. Lowering for codegen will translate streaming channels to single-value channels communicating to/from generated FIFOs.

As part of this change Channel::kind() is replaced with Channel::supported_ops() to more accurately reflect its meaning and to avoid conflation with streaming/single-value channel types.

PiperOrigin-RevId: 341306209