[DSLX] Support fully qualified enum references.
Before modules were introduced, the "kind" of referent was always fully known
at parse time; however, with the introduction of modules, at parse time we no
longer know the kind of referents (via module-local information), and reaching
into dependent modules and referring to their definitions directly is
relatively coupled and potentially error prone.

So instead, we now fall back to less-invariants-known AST structures at parse
time, with more semantic checking afterwards (i.e. in the type checking phase);
whereas before we had ModRef and EnumRef that knew about their associated
(module-local) Import and EnumDef definers explicitly, with the addition of the
new ColonRef we subsume both of those use cases and allow for referring to
EnumDefs, or TypeDefs, present in other modules; i.e.:

  some_module::SomeDef::VALUE

Parses as:

  ColonRef(subject: ColonRef(subject: NameRef(some_module),
			     attr: SomeDef)
	   attr: VALUE)

And we don't need to resolve the SomeDef enum definition in the imported
module. This more-generic less-invariants-known construct can consume ModRef
and EnumRef in the next change since it has all the expressive power.

To determine the construct defining e.g. some_module in the above, we provide
"definer" links on NameDefs to indicate the "containing" AST node causing the
name definition; i.e. if we had the some_module NameRef above:

  name_ref.name_def.definer

would produce the Import node that caused some_module to be defined.

PiperOrigin-RevId: 343927578