[DSLX] Support genrule-made files in genrules wrapping `py_binary`s.
Fixes https://github.com/google/xls/issues/231

Add a test that shows generated file being used in a dslx_test rule.

This shows it being used in both as a dependency and primary-source.

We add additional paths from the command line with a "dslx_path" flag, which is
necessary when Python binaries (which traditionally get their .x files via
runfiles when run directly) are invoked via genrules (where the sources are
placed elsewhere, not in the py binary's runfiles).

PiperOrigin-RevId: 347906827