[fuzzer] Replace Python AST generator with C++ implementation.
Note that this deletes the golden-reference files for particular seeds because
C++ libraries we generally use (absl, random from the standard library) don't
offer the ability to get stable output sequences -- either not at all (absl,
per
https://abseil.io/docs/cpp/guides/random#i-need-my-variate-sequences-to-be-the-same-every-time)
or across conforming implementations (random from the standard library, for
distribution implementations). We're already opting to use random over absl
at the moment because of the seedable API.

One option would be to make a record-and-replay wrapper around the RNG
interface we use to ensure we can replay across platforms (i.e. compilers/libc
implementations), but this seemed like unnecessary complexity when the upside
is just golden reference files that show what kind of output the fuzzer tends
to produce that we can check against. We already have repeatability tests for
repeatability within a given process, and we could add one for repeatability
across process invocations on a given platform (with a given seed) if that was
particularly interesting.

PiperOrigin-RevId: 356327435