Lower XLS IR assert to SystemVerilog assert. An XLS IR assert lowers (at least in the proc generator) to:
  assert($isunknown(cond) || cond) else $fatal("Error message");

The $isunknown is required to avoid spuriously triggering the assert in our simulation infrastructure due to initial X conditions. Later changes will add support for customizable forms of asserts such as project-specific macros.

Also add support for asserts in the simulation infrastructure. An absl::AbortedError is raise if an assert fails.

PiperOrigin-RevId: 356761569