[DSLX] Add support for parsing token types.
Remains to be seen whether we can avoid users having to do token threading for
sequencing while retaining all the control in an intuitive way, but exposing it
allows us to make forward progress as we explore the right sugar to lay on top.

We'll plumb tokens through as part of the "core language" that maybe users will
generally use some decorated or fancified form of.

PiperOrigin-RevId: 356795310