[DSLX] Pre-factor: make fn being typechecked available via inference context.
This will be necessary for determining at typechecking time whether a
callee requires an implicit token/bool for propagating fail!(), and
transitively propagating that property upwards.

Also it's just kind of better to have the function pointer in the
FnStackEntry instead of name (string) / module combo, it is the more
precise information about what's being type checked.

Eventually we probably want to do a pre-pass to determine the
typechecking order, instead of discovering it dynamically and
noting what we're doing in this "fn_stack" entity (i.e.
eventually I want to rip out the fn_stack entirely).

PiperOrigin-RevId: 359145428