[DSLX] Tighten up invariants around TypeInfo.
This was painful, but I think quite worth it! With this change the invariants
are much more clear.

Before, `TypeInfo`s got "smashed together" via this `Update()` call, so you
could potentially (and did in practice) have AST nodes from imported modules
available in the maps of the importing modules. So you would end up looking up
AST nodes from imports in the importer type info.

Now, AST nodes are only looked up in their relevant `TypeInfo` tree. Every
module has a *root* `TypeInfo`, which holds type information that is
independent of any particular parameterized instantiation. Instantiations have
the root type info as their parent, and the parent maps to them via the
instantiation map, which is keyed off of `Invocation` nodes, which drive
parametric instantiations. That root `TypeInfo` also stores the import map for
the module, which in turns points at imported-module root `TypeInfo`s, but
there is no mixing of AST nodes.

The `ImportCache` can be used to resolve the root `TypeInfo` for any module.

So now AST nodes can only get looked up in their module's `TypeInfo` tree. This
means that the "execution engines" (abstract and concrete) -- namely, deductive
type inference, interpretation, or IR conversion -- have to set up the
appropriate `TypeInfo` before traversing into imported modules and acting on
the AST nodes from that scope. For this purpose we use RAII
`ScopedTypeInfoSwap` classes, which establish the `TypeInfo` we're about to
start working with and restore it back to its original value when complete.

Some of the uses of "module" are clarified to mean "entry module" -- this
clarifies when those variables don't change as we traverse into executing
within imported modules.

PiperOrigin-RevId: 360213557