[DSLX] Switch top-level-bindings for module to live on ImportCache.
This a) avoids us recalculating the top level bindings needlessly, and b) gives
a central resource in the future where we can track whether we're re-entering
module evaluation, even beyond interpreter scope.

This will pave the way for typechecking to interpret constexpr expressions
(which spawns a new interpreter for each expression) in a collaborative way
with the interpretation. Currently, the typechecker can request the interpreter
evaluate an expression, the interpreter will go to evaluate the module top
level, and it will find type information is missing at the top level beyond
what the typechecker has gotten to!

Keeping information about work-in-progress of the module's top level at the
ImportCache scope (as noted above) will let us resolve this, so that the
typechecker and the interpreter can both scribble and pick up notes on progress
in a shared data structure.

So, summarily, this is a bit of a pre-factoring for fixing an issue, but it
makes things better nonetheless by not having to recompute the top level so
many times.

Also:

* Makes sure the parent link in `InterpBindings` is const and that the top
  level bindings are only given back as const to avoid accidentally mutating
  them.
* One test needlessly imported std.
* One test file was ophaned, not used in any corresponding dslx_test, fixed
  that (and another was useless, just deleted that one).

PiperOrigin-RevId: 360718574