[DSLX] Fix constexpr eval via shared import progress btw typecheck & interp.
Add a test that demonstrates TypeInfo XLS_RET_CHECK assertion.

When constexpr evaluation happens within a module, we currently make a call to
InterpretExprAsInt from the type inferencing process -- this causes us create
top-level bindings for the module that expression is defined inside. However,
the interpreter evaluated everything at module scope (it didn't know about the
work-in-progress point for the typechecker), so it could encounter an
expression it wanted type information for, where none was available yet.

Having the typechecker note what the "work in progress" point is, the
interpreter knows when to break out of module-level evaluation, and we don't
get the error any longer.

PiperOrigin-RevId: 360968289