[DSLX] Add test of for loop ranging up to constexpr with cast.
Provides "parametric builtin instantiation" with a callback so that they can
request constexpr evaluation of their arguments. This is first used in the
special range() callback per the reasoning in the GitHub issue noted below, but
range() builtin will really be eliminated as it is in the longer term in favor
of range syntax.

With the ability to request constexpr evaluation we also place a constexpr map
onto the TypeInfo structure -- which has its tree formation used for parametric
instantiations -- so requesting constexpr evaluation of a parametric binding
works as we expect (it goes into the corresponding TypeInfo for that parametric
instantiation).

Fixes google/xls#305

PiperOrigin-RevId: 361172618