Add missing transaction rollback - in Parser::ParseComparisonExpression(), the transaction could be destroyed without completion (causing a CHECK-fail) if BinopKindFromString() failed. Adding a Cleanup ensures "completion" on such failure.
In practice, though, such a failure is not possible - the earlier "PeekTokenIn(kComparisonKinds)" guarantees that "op" is a recognized binop. That being said, it's a good idea to use correct patterns, and without a cleanup, the previous transaction pattern was incorrect. Since this failure isn't possible, a unit test isn't possible, either.

PiperOrigin-RevId: 362303793