Fix for some of the OSS fuzz test breakage.
A few tweaks and checks to ensure we don't generate nil tuples.

Under some scenarios we could end up with empty tuples, even when the
"generate_empty_tuples" is false, which OSS fuzzing was hitting.
It was generally when we would exceed the max bit count requested,
where we'd drop everything from the suggested retval tuple count.

Indicates maybe we should make more fuzz targets for coverage
so chances of /only/ finding something in OSS build is reduced?

Note this still breaks even with this change due to
https://github.com/google/xls/issues/350

PiperOrigin-RevId: 364360858