[DSLX] Refactoring in the conversion order extractor and more test cases.
* Notes which constant definitions are local in the AST.

Some tweaks:

* Factors out the handling of map for instantiation.
* Rename Invocation to Instantiation in the TypeInfo.
* Validate parametrics when Callee records are constructed in
  extract_conversion_order.

A few new test cases:

* Test case where parametric and local const are used in match.
* Test for a local const value in a parametric function.

The most interesting thing encountered in the course of this refactoring was
the following scenario: most builtins are parametric, but map() is higher
order, so it's parametric in the type of the mapped function. The function
being mapped, however, may be (but is not necessarily) parametric. So the
representationally interesting question is where to hang the parametric
bindings in the type information for the mapped function, since usually we
associate them with the instantiation of map().

When the mapped function is parametric, but builtin, we want to note the
parametric bindings (for use in running the builtin), but we don't have
user-defined-function TypeInfo.  What his change does (as noted in the TypeInfo
comments) is uses nullptr for this scenario (again, the
mapping-a-parametric-builtin).

PiperOrigin-RevId: 364910646