Enabling parametric values to be InterpValues instead of only int64_t.
This is the first stage - plumbing through InterpValue support everywhere. Next, wherever possible, we should remove explicit int64_t from the supported set (as in ConcreteTypeDim::Variant), as A) InterpValue is strictly a superset of int64_t and B) the existence of two overlapping types is a very jagged-edged special case (as attested by _much_ recent debugging).

Resolves #343.

PiperOrigin-RevId: 365141648