[DSLX] Factor out "runner" functionality from main tool to library.
Also moves quickcheck JIT functionality out of ir_jit translation unit (where
it was not supposed to be long term) to colocate it with these runner routines.
Could probably be factored out into a dedicated quickcheck module but this
seemed ok as a single refactoring step for now.

This allows us to test that JIT invocations are happening more easily /
programmatically.

Note that: 1) the python bindings for this were totally unused now, 2) they
assumed the quickcheck functionality was co-located with the ir_jit translation
unit and didn't take a jit cache.

(Our dead code detection unfortunately can't see through bindings, only within
C++ land, so I was unaware of the dangling uselessness in 1.)

With the quickcheck functionality moved (so it could see the jit cache
definition) it wasn't in the right place per (2) and combined with (1) it was
kind of "guess I'll take the opportunity to delete this since it's not even
used anymore".

PiperOrigin-RevId: 366152833