Adds DSLX implementation of a 32-bit FMA unit.
This implementation is straightforward, given an understanding of FP add and mul - the only difficulty was determining the correct number of significand precision bits (48 for mul, 72 for add) and their handling (whether or not to preserve an implicit 1 bit or which bits to use for rounding).

This has been tested against glibc/libm 2.31, via fma_32_test, over several hundred billion randomly-generated samples to produce bit-accurate results.

Future work will generalize this code in terms of APFloat.

PiperOrigin-RevId: 367261342