[DSLX] Change trace output format via command line flag to interpreter_main.
This is a quick and easy-to-implement stopgap until we have a `trace_fmt!()`
builtin that can do arbitrary formatted logging.

Dependency injects the preference down into the interpreter, and the builtin
can retrieve the preference from the AbstractInterpreter interface.

Example:

```
bazel test -c opt //xls/dslx/tests:trace_dslx_test \
  --test_arg=--trace_format_preference=binary
```

Also bundles up the `ParseAndTest()` optional arguments into a
`ParseAndTestOptions` struct, since it was getting to be an unweildy number of
optional arguments with the addition of this new `trace_format_preference`.

Towards google/xls#372

PiperOrigin-RevId: 368741825