[DSLX] Fix for a "last expression" bug.
Eliminates the need to track any kind of notion like that -- simply use the
BValue that results from the body of the function. This is why expressions are
great, not sure what I was thinking in the first place. :-) I think this was a
vestige from before we allowed `param()` nodes in retval position (without the
`identity()` function).

Also adds checks that converted IR values match types from type-inference
types.

Along the way found that for a one-arm match we'd make a 0-operand concat,
which blew up in one of the ConcatSimplificationPass transforms, so fixes to
permit it, but it still doesn't pass muster in the Z3 IR equivalence check
since Z3 doesn't like zero bit values, filed google/xls#400 for that.

PiperOrigin-RevId: 369507880