[DSLX] Complete migration of IR converter tests to use testdata files.
Because many of the implementation details of the (unoptimized) IR conversion
don't matter, e.g. how many nodes are spit out, whether values are wrapped with
identity nodes, etc. it makes sense to use out of line files for these, so bite
the bullet and move them all instead of piecemeal like I was doing before. :-)

PiperOrigin-RevId: 370157078