[DSLX] Make functions with `fail!()` capable of emitting the `assert` IR.
This effectively adds a tiny little effect inference system to the typechecker
-- if the `fail!` builtin is invoked, transitively the caller of the fallible
function is also fallible, and so on. This is tracked via
`TypeInfo::*RequiresImplicitToken` which indicates that the "implicit token"
calling convention is necessary.

"Implicit token" calling convention effectively prepends `(seq: token, activated: bool)` as
parameters and wraps up the return type to thread the post-assert token out as
`(token, OriginalReturnT)`. More extensive markdown on this to follow (needs to
be converted from Google Doc).

For each control-looking structure in the DSL, a control predicate is able to
be computed, which can guard implicitly effecting ops like `fail()` when it is
converted to a (predicated) assert IR op. Control predicate construction is
enabled for all control constructs in this CL, and tested via IR conversion
tests (as well as by trying to flip the flag default to true by hand and making
sure nothing broke above the pipeline generator).

To turn this on by default we need a follow-on change that exposes *entry
point* routines *without* the "implicit token" calling convention, so they can
be JIT invoked / pipeline-generated like before.

Updates the JIT wrapper generator to detect the "implicit token" calling
convention and expose the underlying "original" signature -- we may be able to
eliminate this in the future if we want to guarantee that *only* entry
functions are elligible to be JIT compiled.

Also, more minor:

* Add Assert/AfterAll to delay models as zero cost.
* Package is decorated with an `ir_to_dslx` function mapping during conversion
  called `PackageData` so we can reverse-lookup which DSLX function an IR
  function originated from (e.g. so we can determine if it wants implicit token
  convention in the DSLX type info).

Towards google/xls#232

PiperOrigin-RevId: 371226265