[build_rules] Mangling macro, avoids users typing mangled names in build files.
When the new Bazel rules (that use depsets) are enabled by default we will
think more deeply about how to make this (and the linkage model in general)
more clean and do it all under the hood -- at that point we'll have real
modules to assume instead of smashing all the source files into a single
package. i.e. This is just a stopgap we should revisit, along with linkage in a
broader sense, when we cut over.

PiperOrigin-RevId: 373616433