[DSLX] Allow for loop bindings' type annotation to be optionally specified.
Resolves some old TODOs.

### On the dslx_builtin type inference change

We needed the type of the `range()` builtin to properly say what type it was
spitting out (an array of some-width type), whereas it was previously saying it
produced unit type `()` -- by spitting out the correct array type we can grab
the element type of that iterable expression and use that when the
user-supplied type annotation is missing; e.g. when we do:

```
for (i, accum) in range(u7:0, u7:4) {

}(...)
```

How do we determine what the type of `i` is? It's a u7, but unless we can
interrogate the value `range()` spits out (to note it's a `u7[4]`) we won't be
able to tell, now that users are not giving us explicit annotations. So, we
need `range()` to really tell us its resulting array type instead of being
fake.

The original reason it was fake was because we didn't have the ability to find
what the constexpr values were that were passed to range, but now that we do
(yay!), we can use that facility to grab the constant values being passed to
determine the resulting type.

PiperOrigin-RevId: 374235911