Fixes repeated failure messages for failing DSLX tests.
Prior to this change, a failed assert in a test function would cause N+1 copies of the failure message to print ("Failure Error: <span>: The program being interpreted failed! ..."). This was due to the fact that the N'th let in a function is evaluated as part of the N-1'th let, which is evaluated as part of the N-2'th let, so on and so forth. Each of those evaluations ends up in Interpreter::Evaluate(), which sees if that let's evaluation failed, and prints an error in that case. This failure status was propagated up the call stack, meaning that a failure in one let would hit the failure-printing code in each precending let.

This fixes that by moving printing to evaluation root, i.e., one function above the first call to Interpreter::Evaluate().

PiperOrigin-RevId: 374665328