[interp] Remove stats object from interpreter. Had no line coverage.
I originally added it as a way of collecting coverage information, but seems
like it was a mis-step since it's not really used and now we'd want the JIT to
collection this info -- if we had a coherent way of using it more we'd maybe
want to redo it it a bit in any case.

Also notable, the fuzzer collects static information on the varieties of
operations emitted (though it does not collect dynamic information like
overshift amounts or runtime observed values).

PiperOrigin-RevId: 374791581