Adds literal string support to DSLX.
Internally, strings are just held as sequences of u8s; little additional handling is done on top of that.

It's possible that we may want to treat strings specially, so that the ultimate Verilog can contain human-readable strings, in which case we'll need to add special string support in the IR, but that will require further thought before implementation.

PiperOrigin-RevId: 374976032