[docs] Add documentation scrape test.
Fairly simple script that scrapes code blocks out of the documentation and runs
them through the DSL interpreter main binary -- already found a bunch of places
there were inconsistencies!

Should generally help make sure our docs are more up to date and sound.

For future work, for DSLX "snippets" (things that are not complete code blocks
that are displayed, but that are used to show DSL constructs), we may want to
define a "hidden" prelude before the snippet that makes the block executable,
and have some way to delimit what subset of the snippet is shown when rendered.

Fixes google/xls#378

PiperOrigin-RevId: 375183932