[DSLX] Wrap public fallible functions so they can serve as entry points.
When fallible functions are converted they use the implicit token calling
convention, which would look different from what the user wrote in the DSL code
when exposed to the outside world via IR JIT or verilog module interfaces.

By creating a wrapper that serves as an entry point, we can keep the signature
-- that would be called by the outside world -- the same for fallible
functions as infallible ones.

If there is no `pub` keyword on the function, and it is not the IR package's
entry point, we assume it is there for internal call graph purposes, and don't
create a wrapper. Creating a wrapper for every fallible function seemed
excessive.

This will potentially be revisited to put a notion of "possible entry point"
through the whole stack, but for now, this works with the grain of how IR
packages are currently exposed and optimized.

PiperOrigin-RevId: 375190654