[public] Make a "runtime build action" API for proto to DSLX conversion.
Subprocess invocation of the proto_to_dslx binary is not a viable option in
some environments, and it's the kind of thing we can-and-likely-will have a
Bazel build rule for, so it fits as a public API.

Note that this only exposes the "single protobuf definition file with no
imports" functionality, because making fake filesystems seems a lot to export
via a public API if it's largely YAGNI at this point.

Aside: the sinking of some of the FATALs done in this CL squashes a few
over-eager gcc warnings while we were in here.
PiperOrigin-RevId: 376022312