[DSLX] Refactor ParametricInstantiator arg types / spans to AoS instead of SoA.
Reads more naturally in C++ and ensures types and spans are paired with equal
sizes.

The only (minor) wrinkle that presents itself at some points in the code is
that sometimes the types have been locally deduced as unique_ptrs, and so must
be owned for the duration of the instantiate call, independently of the
InstantiateArgs (which have unowned references).

PiperOrigin-RevId: 376188898