[DSLX] Remove signed comparison builtins, move to stdlib.
Back in the primordial days there was no signed type, so there was no way to
write a signed comparison in userspace. Now that there is, these don't need to
be builtins, and they only serve as slightly convenience helpers, so we move
them to the stdlib and out of builtin space.

PiperOrigin-RevId: 376895928