Swap endianness of array representation in Verilog. When an XLS array is flattened in Verilog, previously the zero-th array element was at the highest bit index in the flattened bit vector. With this change the zero-th array element is at the lowest bit index. This comes into play as arrays are flattened when passed in or out of module ports so this change is visible as it affects the interface of the generated Verilog modules.
For example, consider an array 'a' of type bits[32][3]. Before this change, 'a' was flattened like so (in Verilog):

  a_flattened = {a[0], a[1], a[2]};

With this change:

  a_flattened = {a[2], a[1], a[0]};

The motivation for this change is that this new endianness corresponds to the endianness of SystemVerilog packed arrays. At some point XLS will switch to using packed arrays as the representation of arrays in SystemVerilog.

PiperOrigin-RevId: 377315024