Add ports to block declaration in the IR. Previously blocks were represented solely as input_port and output_port nodes in the block body. Example:
  block my_block {
    foo: bits[32] = input_port(id=1)
    ...
    out: bits[32] = output_port(bar, id=42)
  }

This CL adds a ordered list of ports similar to function parameters, for example:

  block my_block(foo: bits[32], out: bits[32]) {
    foo: bits[32] = input_port(id=1)
    ...
    out: bits[32] = output_port(bar, id=42)
  }

The ordered list carries the port order as emitted in the Verilog and also improves (human) readability by making the interface more explicit. There is some duplication in this representation as ports are represented in the port order and as input_port and output_port nodes, but this compromise is necessary as the nodes (input_port/output_port) carry important information (the node id).

Also, add a clock port representation to blocks. Clock is not treated as a normal signal but rather as a special sentinel value on the list of ports which carries its port position and name. Example text representation:

  block my_block(clk: clock, foo: bits[32], out: bits[32]) {
    foo: bits[32] = input_port(id=1)
    ...
  }

PiperOrigin-RevId: 377974350