Fixing occasional crash when crashing fail! or cover! to IR.
WrapEntryIfImplicitToken() checked to see if the entry function to the IR package needed to be token-wrapped, i.e., if it contained a fail! or cover! operation that needed an input & output token. Unfortunately, an IR Package doesn't have a conception of public or private, so it could choose an IR function corresponding to a non-"public" DSLX function as its entry point.

In that case, the IR converter would crash because it does not generate wrappers for non-public functions.

PiperOrigin-RevId: 378159657