Adding builder class for Testbench.
In order to simplify and unify testing across XLS, it'd be helpful to further
simplify test creation and execution, with something like RapidCheck as an
ideal*. The first step in doing this is to de-JIT-ify the Testbench and to
simplify its creation via a type-inferencing builder: TestbenchBuilder can,
in some cases, infer proper behavior based on argument types, e.g., it can
compare two ints or floats (with minor additional logic), or it can randomly
populate a tuple of tuples of ... built-in types.

Follow-ons to this change will include google test macros (for simpler unit
tests) and the addition of more defaults, either in supported types
(e.g., std::vector<X>) or in behavior (e.g., generate serial inputs instead
of uniform-random ones).

* RapidCheck itself is a difficult fit for the things we often do, which is why
this isn't a CL adding it as an XLS dependency.

PiperOrigin-RevId: 378242990