Plumbing support for coverpoints through codegen.
With this change, DSLX `cover!` statements will be present as `cover property` statements in the emitted SystemVerilog (they will be dropped if outputting as Verilog). Verilog cover statements can only be evaluated in the context of a clock, so they will only be emitted by the pipeline generator and not the combinational one.

Resolves issue #436.

PiperOrigin-RevId: 378492761